+++ src/util/SqlModel.js
@@ -38,7 +38,7 @@ module.exports = class SqlModel {
     this.state = 0;
   }
 
-  queue(action) {
+  enqueue(action) {
     const self = this;
     return new Promise((resolve, reject) => {
       if (self.state !== 1) {
@@ -54,7 +54,7 @@ module.exports = class SqlModel {
     function action() {
       return this.data;
     }
-    return this.queue(action);
+    return this.enqueue(action);
   }
 
   fetch() {
@@ -95,7 +95,7 @@ module.exports = class SqlModel {
       }
       return keys;
     }
-    return this.queue(action);
+    return this.enqueue(action);
   }
 
   filter(query, value) {
@@ -122,7 +122,7 @@ module.exports = class SqlModel {
       }
       return undefined;
     }
-    return this.queue(action);
+    return this.enqueue(action);
   }
 
   findOne(query, value) {
@@ -202,6 +202,17 @@ module.exports = class SqlModel {
             return resolve(document);
           }
           resolve(undefined);
+        })
+        .catch(reject);
+    });
+  }
+
+  deleteMany(query, value) {
+    return new Promise((resolve, reject) => {
+      this.filterKeys(query, value)
+        .then(keys => {
+          const deleted = keys.map(key => this.deleteOne({ _id: key }));
+          resolve(Promise.all(deleted));
         })
         .catch(reject);
     });